server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
